# api/openapi.yaml
openapi: 3.1.0
info:
  title: NimbusDB API
  version: 1.0.0-phase1
  description: |
    Control-plane REST API for the NimbusDB serverless PostgreSQL platform.
    Spec-first contract (ADR-012): handlers and clients are generated/validated from this file.
    Phase 1 surface: bootstrap, orgs, members, API keys, project records, audit log.
    Data-plane resources (branches, endpoints, backups, imports) arrive in Phase 2+ per
    docs/architecture/API_SPECIFICATION.md.
servers:
  - url: https://api.db.nimbus.app/v1
  - url: http://localhost:8080/v1
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /projects/{prj}/branches:
    parameters:
      - name: prj
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [branches]
      operationId: listBranches
      summary: List branches of a project
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: Branches (endpoints omitted; fetch a branch for them)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Branch" }
                  next_cursor: { type: [string, "null"] }
                required: [data]
        "404": { $ref: "#/components/responses/Problem" }
    post:
      tags: [branches]
      operationId: createBranch
      summary: Create a branch
      description: |
        Creates a branch record with its rw_direct + rw_pooled endpoints in
        `provisioning` state. `from_branch` defaults to the project's default
        branch. Point-in-time sources (`at`) arrive in Phase 4.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 63 }
                from_branch: { type: [string, "null"] }
                role: { $ref: "#/components/schemas/BranchRole" }
                compute_min_cu: { type: number, minimum: 0.25, maximum: 8 }
                compute_max_cu: { type: number, minimum: 0.25, maximum: 8 }
                suspend_timeout_s: { type: integer, minimum: 0, maximum: 86400 }
              required: [name]
      responses:
        "201":
          description: Created branch (with endpoints)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /branches/{br}:
    parameters:
      - name: br
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [branches]
      operationId: getBranch
      summary: Get a branch with its endpoints
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [branches]
      operationId: updateBranch
      summary: Update branch settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 63 }
                compute_min_cu: { type: number, minimum: 0.25, maximum: 8 }
                compute_max_cu: { type: number, minimum: 0.25, maximum: 8 }
                suspend_timeout_s: { type: integer, minimum: 0, maximum: 86400 }
                retention_days: { type: integer, minimum: 1, maximum: 30 }
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Branch" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [branches]
      operationId: deleteBranch
      summary: Delete a branch
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Problem" }
        "409":
          description: Default branch is delete-protected
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /branches/{br}/endpoints:
    parameters:
      - name: br
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [branches]
      operationId: listEndpoints
      summary: List a branch's endpoints
      responses:
        "200":
          description: Endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Endpoint" }
                required: [data]
        "404": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: "ndb_<64 hex chars>"
  parameters:
...
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        request_id: { type: string }
      required: [title, status]
...
    BranchRole:
      type: string
      enum: [production, preview, development]
    ResourceState:
      type: string
      enum: [provisioning, ready, suspended, error, deleting]
    Compute:
      type: object
      properties:
        min_cu: { type: number }
        max_cu: { type: number }
        suspend_timeout_s: { type: integer }
      required: [min_cu, max_cu, suspend_timeout_s]
    Branch:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        parent: { type: [string, "null"] }
        name: { type: string }
        role: { $ref: "#/components/schemas/BranchRole" }
        state: { $ref: "#/components/schemas/ResourceState" }
        compute: { $ref: "#/components/schemas/Compute" }
        retention_days: { type: integer }
        created_at: { type: string, format: date-time }
        endpoints:
          type: array
          description: Present on single-branch reads and creation responses.
          items: { $ref: "#/components/schemas/Endpoint" }
      required: [id, project_id, parent, name, role, state, compute, retention_days, created_at]
    Endpoint:
      type: object
      properties:
        id: { type: string }
        branch_id: { type: string }
        kind:
          type: string
          enum: [rw_direct, rw_pooled, ro_pooled]
        host: { type: string }
        state: { $ref: "#/components/schemas/ResourceState" }
        created_at: { type: string, format: date-time }
      required: [id, branch_id, kind, host, state, created_at]
...